+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.4.0 Under development
 
 - Bug #37: Move `Message` enum to `Helper\Exception` namespace and clean up unused cases (@terabytesoftw)
+- Bug #38: Update method calls to use `getAttributes()` method for consistency in `BaseBlock`, `BaseInline`, and `BaseVoid` classes (@terabytesoftw)
 
 ## 0.3.1 December 26, 2025